Upgrade to tilt 2.x
Some of the code in tilt 1.4 generates warnings when run under Ruby 2.4.

For example:

    /path/to/.gem/ruby/2.4.2/gems/tilt-1.4.1/lib/tilt/haml.rb:55:in
    `block in precompiled_postamble':
    Haml::Engine#precompiled_method_return_value at
    /path/to/.rubies/ruby-2.4.2/lib/ruby/2.4.0/forwardable.rb:157
    forwarding to private method
    Haml::Compiler#precompiled_method_return_value

Upgrading to tilt 2.0 fixes the problem.

tilt 2.0 automatically requires gems that are used to process templates
(e.g. `maruku`), which tilt 1.4 didn't do. I've therefore had to update
the code that registers the Markdown template handlers in
`lib/nesta/models.rb`.

We ignore `LoadError` as we don't require any of these handlers to be
present, as `rdiscount` is used by default. It might be sensible to stop
requiring `rdiscount` here, and allowing each site to include whichever
Markdown handler is preferred in its `Gemfile`, but that's out of scope
of this commmit.

Were we to move to specifying the default Markdown processor in each
site's `Gemfile` I suspect it would also make sense to check whether
Haml's inline Markdown feature works with any Markdown processor, and
update the docs accordingly if it does. I suspect that Haml's inline
Markdown processing might be one of the reasons why `rdiscount` is hard
coded as a default.